tests: add script to check that stanc3 recognizes entry points
For each Torsten entry point exposed via stanc3, it'd be good to have
an indication that it's wired up correctly.  This is already the case
for entry points that are covered in example-models/.  However, it
probably isn't worth adding and maintaining an example for every case
just to check that all entry points are recognized.

Instead add a script that creates a non-functional dummy model for
each entry point and verifies that building the model does not fail
with the "undeclared identifier" error that stanc3 would emit if it
didn't recognize the entry point at all.

The limitations of this script are that

  1) it relies on stanc3 emitting a particular error message, so a
     stanc3 update could lead to it silently passing

  2) similarly, it relies on stanc3 not signaling some an unrelated
     error _before_ the undeclared identifier error

  3) it requires keeping the list of entry points in the script up to
     date.  That's not so bad when removing/renaming an entry point,
     because the script will fail loudly, but there is a risk of
     forgetting to add a new entry point.